Safari-611.1.10/LayoutTests/accessibility/mac/selection-initial-expected.txt

This tests that initial selection is set when arrow keys are pressed and no selection exists.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isSelection() is false
PASS isSelection() is false
PASS accessibilityController.enhancedAccessibilityEnabled is true
PASS isSelection() is false
PASS isSelectionAtStartOfDocument() is true
PASS isSelection() is false
PASS isSelectionAtEndOfDocument() is true
PASS isSelection() is false
PASS isSelectionAtStartOfDocument() is true
PASS isSelection() is false
PASS isSelectionAtEndOfDocument() is true
PASS isSelection() is false
PASS isSelection() is false
PASS isSelection() is false
PASS isSelectionAfterButton() is true
PASS isSelection() is false
PASS isSelectionBeforeButton() is true
PASS isSelection() is false
PASS isSelectionAfterButton() is true
PASS isSelection() is false
PASS isSelectionBeforeButton() is true
PASS successfullyParsed is true

TEST COMPLETE
